% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{acronyms}
\alias{acronyms}
\title{Acronyms}
\format{
A data frame with the acronyms of 16823 diatoms
}
\usage{
data(acronyms)
}
\description{
Contains the acronyms for diatom species
}
\keyword{datasets}
