% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{cemfgs_rb}
\alias{cemfgs_rb}
\title{CEMFGS_RB}
\format{
A data frame with the ecological values for 495 species
}
\source{
\doi{10.1016/j.ecolind.2017.07.007}
}
\usage{
data(cemfgs_rb)
}
\description{
Index values for diatom species combining their ecological guilds with their size classes
}
\references{
B-Béres, V., Török, P., Kókai, Z., Lukács, Á., Enikő, T., Tóthmérész, B., & Bácsi, I. (2017). Ecological background of diatom functional groups: Comparability of classification systems. Ecological Indicators, 82, 183-188.
}
\keyword{datasets}
