% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{dbc_offline}
\alias{dbc_offline}
\title{DBC (offline)}
\format{
A data frame with ecological and morphological information for 8066 diatoms
}
\source{
\url{https://www.kmae-journal.org/articles/kmae/abs/2012/03/kmae120025/kmae120025.html}
}
\usage{
data(dbc_offline)
}
\description{
Diatom database from the 'Diat.Barcode' project V12.1
}
\references{
Rimet F. & Bouchez A., 2012. Life-forms, cell-sizes and ecological guilds of diatoms in European rivers. Knowledge and management of aquatic ecosystems, 406: 1-14. DOI:10.1051/kmae/2012018
}
\keyword{datasets}
