% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ddi}
\alias{ddi}
\title{DDI}
\format{
A data frame with the ecological values for 137 species
}
\source{
\url{https://link.springer.com/article/10.1007/s10661-012-2607-z}
}
\usage{
data(ddi)
}
\description{
Index values for diatom species included in the DDI index
}
\references{
Álvarez-Blanco, I., Blanco, S., Cejudo-Figueiras, C., & Bécares, E. (2013). The Duero Diatom Index (DDI) for river water quality assessment in NW Spain: design and validation. Environmental monitoring and assessment, 185, 969-981.
}
\keyword{datasets}
