% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{des}
\alias{des}
\title{DES}
\format{
A data frame with the ecological values for 622 species
}
\source{
\url{https://pascal-francis.inist.fr/vibad/index.php?action=getRecordDetail&idt=PASCAL8060205402}
}
\usage{
data(des)
}
\description{
Index values for diatom species included in the DES index
}
\references{
Descy, J. P. 1979. A new approach to water qualityestimation  using  diatom.  Beih.  Nov  Hedw. 64:305-323
}
\keyword{datasets}
