% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diat_checkName.R
\encoding{UTF-8}
\name{diat_checkName}
\alias{diat_checkName}
\title{Searches all the taxa database for the input name}
\usage{
diat_checkName(taxaname, byword = F)
}
\arguments{
\item{taxaname}{the name of the taxa (genus, species, variety) to be checked against the internal DB}

\item{byword}{if byword = F (default), the input string will be searched without splitting words. If True, each word will be searched separately}
}
\description{
Searches all the taxa database for the input name, returns a list with the results
}
\keyword{bioindicator}
\keyword{biotic}
\keyword{diatom}
\keyword{ecology}
