% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diat_diversity.R
\encoding{UTF-8}
\name{diat_diversity}
\alias{diat_diversity}
\title{Calculate diversity parameters for diatoms using the vegan package}
\usage{
diat_diversity(resultLoad)
}
\arguments{
\item{resultLoad}{The resulting list obtained from the diat_loadData() function}
}
\description{
The input for these functions is the resulting dataframe obtained from the diat_loadData() function, to calculate diversity data using the vegan package
Sample data in the examples is taken from:
\itemize{
\item Nicolosi Gelis, María Mercedes; Cochero, Joaquín; Donadelli, Jorge; Gómez, Nora. 2020. "Exploring the use of nuclear alterations, motility and ecological guilds in epipelic diatoms as biomonitoring tools for water quality improvement in urban impacted lowland streams". Ecological Indicators, 110, 105951. https://doi:10.1016/j.ecolind.2019.105951
}
Diversity index (Shannons H') is calculated using the vegan package, following:
\itemize{
\item Shannon, C. E., and Weaver, W. (1949). ‘The Mathematical Theory of Communication.’ (University of Illinios Press: Urbana, IL, USA.)
}
}
\keyword{bioindicator}
\keyword{biotic}
\keyword{diatom}
\keyword{ecology}
