% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diat_Load.R
\name{diat_findAcronyms}
\alias{diat_findAcronyms}
\title{The function 'loadFile' loads the CSV or dataframe file, sets the Output folder for the package, and conducts both an exact and an heuristic search of the species' names
The function 'findAcronym' conducts both an exact and an heuristic search of the species' names and tries to convert it to its acronym in the internal database}
\usage{
diat_findAcronyms(species_df, maxDistTaxa = 2)
}
\arguments{
\item{species_df}{The data frame with your species data. Species as rows, Sites as columns. If empty, a dialog will prompt for a CSV file}

\item{maxDistTaxa}{Integer. Number of characters that can differ in the species' names when compared to the internal database's name in the heuristic search. Default = 2}

\item{isRelAb}{Boolean. If set to 'TRUE' it means that your species' data is the relative abundance of each species per site. If FALSE, it means that it the data corresponds to absolute densities. Default = FALSE}
}
\description{
The input file for the package is a dataframe or an external CSV file. Species should be listed as rows, with species' names in column 1.
The following columns have to be the abundance of each species (relative or absolute) in each sample (column).
The first row of the file has to contain the headers with the sample names.
If a dataframe is not specified as a parameter (species_df), the package will show a dialog box to search for the CSV file
A second dialog box will help set up an Output folder, where all outputs from the package will be exported to (dataframes, CSV files, plots in PDF)
The package downloads and installs  a wrapper for the Diat.Barcode project. Besides citing this package, the Diat.Barcode project should also be cited if the package is used, as follows:
\itemize{
\item Rimet, Frederic; Gusev, Evgenuy; Kahlert, Maria; Kelly, Martyn; Kulikovskiy, Maxim; Maltsev, Yevhen; Mann, David; Pfannkuchen, Martin; Trobajo, Rosa; Vasselon, Valentin; Zimmermann, Jonas; Bouchez, Agnès, 2018, "Diat.barcode, an open-access barcode library for diatoms", https://doi.org/10.15454/TOMBYZ, Portail Data Inra, V1
}
Sample data is taken from:
\itemize{
\item Frederic Rimet; Philippe Chaumeil; Francois Keck; Lenaïg Kermarrec; Valentin Vasselon; Maria Kahlert; Alain Franc; Agnes Bouchez. 2016. R-Syst::diatom: an open-access and curated barcode database for diatoms and freshwater monitoring. Database, 2016, 1-21.http://database.oxfordjournals.org/content/2016/baw016.full?keytype=ref&ijkey=H324uA95JzzEomz
}
\itemize{
\item Rimet F., Chaumeil P., Keck F., Kermarrec L., Vasselon V., Kahlert M., Franc A., Bouchez A., 2015. R-Syst::diatom: a barcode database for diatoms and freshwater biomonitoring - data sources and curation procedure. INRA Report, 14 pages http://dx.doi.org/10.5281/zenodo.31137
}
}
\examples{
# EXAMPLE 1:
data("environmental_data") #ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
data("species_data")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
# EXAMPLE 2: Loads sample data where species are in absolute densities
data("environmental_data_example2")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
data("species_data_example2")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
# Calculates everything
thispackage::calculate_everything(species_df)
}
