% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diat_lobo.R
\encoding{UTF-8}
\name{diat_lobo}
\alias{diat_lobo}
\title{Calculates the Lobo Index (LOBO)}
\usage{
diat_lobo(resultLoad)
}
\description{
The input for all of these functions is the resulting dataframe (resultLoad) obtained from the diat_loadData() function
A CSV or dataframe cannot be used directly with these functions, they have to be loaded first with the diat_loadData() function
so the acronyms and species' names are recognized
References for the index:
\itemize{
\item Lobo, E. A., Callegaro, V. L. M., & Bender, E. P. (2002). Utilização de algas diatomáceas epilíticas como indicadoras da qualidade da água em rios e arroios da Região Hidrográfica do Guaíba, RS, Brasil. Edunisc.
}
\itemize{
\item Lobo, E. A., Bes, D., Tudesque, L., & Ector, L. (2004). Water quality assessment of the Pardinho River, RS, Brazil, using epilithic diatom assemblages and faecal coliforms as biological indicators. Vie et Milieu, 54(2-3), 115-126.
}

Sample data in the examples is taken from:
\itemize{
\item Nicolosi Gelis, María Mercedes; Cochero, Joaquín; Donadelli, Jorge; Gómez, Nora. 2020. "Exploring the use of nuclear alterations, motility and ecological guilds in epipelic diatoms as biomonitoring tools for water quality improvement in urban impacted lowland streams". Ecological Indicators, 110, 105951. https://doi.org/10.1016/j.ecolind.2019.105951
}
}
\examples{
# First either a) load the CSV file with the diat_loadData() function
inputdata <- diat_loadData() #this, without parameters, will prompt dialog boxes to select the CSV file and the Output folder
# or b) use a Species x sample dataframe already in R
inputdata <- diat_loadData(Rdataframe) #this, without parameters, will prompt a dialog box to select the Output folder
#After waiting for the data to load correctly, you can calculate any of the diatom indices included in the package
#Each function is a separate index. For instance:
diat_lobo(inputdata) #calculates the LOBO index (Lobo et al. 2002; 2004)


# Example using sample data included in the package (sampleData):
data("diat_sampleData")
# First, the diat_Load() function has to be called to read the data
# The data will be stored into a list (loadedData)
# And an output folder will be selected through a dialog box
loadedData <- diat_loadData(diat_sampleData)
# Next, we can run any function in the package to obtain results from that dataframe. Check the console to see if enough species/acronyms were
# automatically recognized from your data, or if you need to correct the input data
loboResults <- diat_lobo(loadedData)

}
\keyword{bioindicator,}
\keyword{biotic}
\keyword{diatom,}
\keyword{ecology,}
\keyword{indices}
