% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diat_Ecology.R
\name{diat_morpho}
\alias{diat_morpho}
\title{These functions calculate morphological and ecological parameters from the input data}
\usage{
diat_morpho(resultLoad)
}
\arguments{
\item{resultLoad}{The result dataframe obtained from the loadFiles function}

\item{vandamReports}{Boolean. If set to 'TRUE' the detailed reports for the Van Dam classifications will be reported in the Output. Default = TRUE}
}
\description{
The input for these functions is the resulting dataframe obtained from the loadFiles function to calculate morphological parameters, diversity values (using the vegan package), ecological guilds and the Van Dam ecological preferences
The morphological data (size classes, chlorophlasts) is obtained from the Diat.Barcode project. Besides citing this package, the Diat.Barcode project should also be cited if the package is used, as follows:
\itemize{
\item Rimet, Frederic; Gusev, Evgenuy; Kahlert, Maria; Kelly, Martyn; Kulikovskiy, Maxim; Maltsev, Yevhen; Mann, David; Pfannkuchen, Martin; Trobajo, Rosa; Vasselon, Valentin; Zimmermann, Jonas; Bouchez, Agnès, 2018, "Diat.barcode, an open-access barcode library for diatoms", https://doi.org/10.15454/TOMBYZ, Portail Data Inra, V1
}
Guild classification is obtained from:
\itemize{
\item CITA GUILDS
}
Van Dam classification is obtained form:
\itemize{
\item CITA VAN DAM
}
}
\examples{
# EXAMPLE 1:
data("environmental_data") #ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
data("species_data")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
# EXAMPLE 2: Loads sample data where species are in absolute densities
data("environmental_data_example2")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
data("species_data_example2")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
# Calculates everything
thispackage::calculate_everything(species_df)
}
