% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diat_taxaList.R
\encoding{UTF-8}
\name{diat_taxaList}
\alias{diat_taxaList}
\title{Creates a single list with taxa names from all indices within DiaThor}
\usage{
diat_taxaList()
}
\description{
Creates a single list with taxa names from all indices within DiaThor
}
\keyword{bioindicator}
\keyword{biotic}
\keyword{diatom}
\keyword{ecology}
