% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{disp}
\alias{disp}
\title{DISP}
\format{
A data frame with the ecological values for 143 species
}
\source{
\doi{10.1016/j.ecolind.2018.07.026}
}
\usage{
data(disp)
}
\description{
Index values for diatom species included in the DISP index
}
\references{
Stenger-Kovács, C., Körmendi, K., Lengyel, E., Abonyi, A., Hajnal, É., Szabó, B., Buczkó, K. & Padisák, J. (2018). Expanding the trait-based concept of benthic diatoms: Development of trait-and species-based indices for conductivity as the master variable of ecological status in continental saline lakes. Ecological Indicators, 95, 63-74.
}
\keyword{datasets}
