% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{edi}
\alias{edi}
\title{EDI}
\format{
A data frame with the ecological values for 147 species
}
\source{
\url{https://pubs.acs.org/doi/abs/10.1021/acsestwater.4c00126}
}
\usage{
data(edi)
}
\description{
Index values for diatom species included in the EDI index
}
\references{
Chamorro, S., Moyón, J., Salazar, J., Chamorro, K., Vicuña, Z., Cordero, P., Bécares, E. & Blanco, S. (2024). Water Quality Monitoring in Ecuadorian Streams Using a New Diatom-Based Index. ACS ES&T Water, 4(9), 3816-3823.
}
\keyword{datasets}
