% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{idap}
\alias{idap}
\title{IDAP}
\format{
A data frame with the ecological values for 194 species
}
\source{
\url{https://link.springer.com/article/10.1007/BF00028033}
}
\usage{
data(idap)
}
\description{
Index values for diatom species included in the IDAP index
}
\references{
Prygiel, J., & Coste, M. (1993). The assessment of water quality in the Artois-Picardie water basin (France) by the use of diatom indices. Hydrobiologia, 269(1), 343-349.
}
\keyword{datasets}
