% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{idp}
\alias{idp}
\title{IDP}
\format{
A data frame with the ecological values for 475 species
}
\source{
\url{https://link.springer.com/article/10.1023/A:1011415209445}
}
\usage{
data(idp)
}
\description{
Index values for diatom species included in the IDP index
}
\references{
Gómez, N., & Licursi, M. (2001). The Pampean Diatom Index (IDP) for assessment of rivers and streams in Argentina. Aquatic Ecology, 35(2), 173-181.
}
\keyword{datasets}
