% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ilm}
\alias{ilm}
\title{ILM}
\format{
A data frame with the ecological values for 667 species
}
\source{
\url{http://www.vliz.be/imisdocs/publications/286641.pdf}
}
\usage{
data(ilm)
}
\description{
Index values for diatom species included in the ILM index
}
\references{
Leclercq, L., & Maquet, B. (1987). Deux nouveaux indices diatomique et de qualité chimique des eaux courantes. Comparaison avec différents indices existants. Cahier de Biology Marine, 28, 303-310
}
\keyword{datasets}
