% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{ips}
\alias{ips}
\title{IPS}
\format{
A data frame with the ecological values for 6881 species
}
\source{
\url{https://www.documentation.eauetbiodiversite.fr/notice/etude-des-methodes-biologiques-d-appreciation-quantitative-de-la-qualite-des-eaux0}
}
\usage{
data(ips)
}
\description{
Index values for diatom species included in the IPS index
}
\references{
Coste, M. (1982). Étude des méthodes biologiques d’appréciation quantitative de la qualité des eaux. Rapport Cemagref QE Lyon-AF Bassin Rhône Méditerranée Corse.
}
\keyword{datasets}
