% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{lobo}
\alias{lobo}
\title{LOBO}
\format{
A data frame with the ecological values for 297 species
}
\usage{
data(lobo)
}
\description{
Index values for diatom species included in the LOBO index
}
\references{
Lobo, E. A., Callegaro, V. L. M., & Bender, E. P. (2002). Utilização de algas diatomáceas epilíticas como indicadoras da qualidade da água em rios e arroios da Região Hidrográfica do Guaíba, RS, Brasil. Edunisc.
}
\keyword{datasets}
