% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{pdise}
\alias{pdise}
\title{PDISE}
\format{
A data frame with the ecological values for 455 species
}
\source{
\url{https://link.springer.com/article/10.1007/s10661-023-11378-4}
}
\usage{
data(pdise)
}
\description{
Index values for diatom species included in the PDIse index
}
\references{
Kahlert, M., Fölster, J., & Tapolczai, K. (2023). No lukewarm diatom communities—the response of freshwater benthic diatoms to phosphorus in streams as basis for a new phosphorus diatom index (PDISE). Environmental Monitoring and Assessment, 195(7), 807
}
\keyword{datasets}
