% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{sla}
\alias{sla}
\title{SLA}
\format{
A data frame with the ecological values for 976 species
}
\source{
\url{https://onlinelibrary.wiley.com/doi/abs/10.1002/aheh.19860140519}
}
\usage{
data(sla)
}
\description{
Index values for diatom species included in the SLA index
}
\references{
Sládecek, V. (1986). Diatoms as indicators of organic pollution. Acta hydrochimica et hydrobiologica, 14(5), 555-566.
}
\keyword{datasets}
