% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{taxaList}
\alias{taxaList}
\title{taxaList}
\format{
A data frame with names of 9806 taxa
}
\usage{
data(taxaList)
}
\description{
List of taxa names used in all indices. Gets updated user-end if a new DBC is found
}
\keyword{datasets}
