% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{tdi}
\alias{tdi}
\title{TDI}
\format{
A data frame with the ecological values for 3445 species
}
\source{
\url{https://link.springer.com/article/10.1007/BF00003802}
}
\usage{
data(tdi)
}
\description{
Index values for diatom species included in the TDI index
}
\references{
Kelly, M. G., & Whitton, B. A. (1995). The trophic diatom index: a new index for monitoring eutrophication in rivers. Journal of Applied Phycology, 7(4), 433-444.
}
\keyword{datasets}
