% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_algaebase_list.R
\encoding{UTF-8}
\name{update_algaebase_list}
\alias{update_algaebase_list}
\title{Updates a list of taxa names against AlgaeBase.org}
\usage{
update_algaebase_list(listToUpdate, overwrite = T)
}
\arguments{
\item{listToUpdate}{the name of the taxa (genus, species, variety) to be checked against the internal DB}

\item{overwrite}{if overwrite = T (default), if a new name is found, it replaces the original one. The corrected list is returned. Otherwise, a new list is returned, with the original taxon and the new name found}
}
\description{
The package is a wrapper for the algaeClassify package from the 'AlgaeBase.org' database. Besides citing the DiaThor package, the Diat.Barcode project should also be cited, as follows:
\itemize{
\item Patil, V., Seltmann, T., Salmaso, N., Anneville, O., Lajeunesse, M., Straile, D. 2019. algaeClassify: Determine Phytoplankton Functional Groups Based on Functional Traits.
}
}
\keyword{bioindicator}
\keyword{biotic}
\keyword{diatom}
\keyword{ecology}
