% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diat_getDiatBarcode.R
\encoding{UTF-8}
\name{diat_getDiatBarcode}
\alias{diat_getDiatBarcode}
\title{Loads the 'Diat.Barcode' database into DiaThor in the correct format}
\usage{
diat_getDiatBarcode()
}
\description{
The package downloads and installs a wrapper for the 'Diat.Barcode' project. Besides citing the DiaThor package, the Diat.Barcode project should also be cited, as follows:
\itemize{
\item Rimet F., Gusev E., Kahlert M., Kelly M., Kulikovskiy M., Maltsev Y., Mann D., Pfannkuchen M., Trobajo R., Vasselon V., Zimmermann J., Bouchez A., 2019. Diat.barcode, an open-access curated barcode library for diatoms. Scientific Reports. https://www.nature.com/articles/s41598-019-51500-6
}
}
\keyword{bioindicator}
\keyword{biotic}
\keyword{diatom}
\keyword{ecology}
