% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{diat_sampleData}
\alias{diat_sampleData}
\title{Sample Data}
\format{
A data frame with the abundance of 164 diatoms in 108 sampled sites
}
\source{
\url{https://www.sciencedirect.com/science/article/abs/pii/S1470160X1930946X}
}
\usage{
data(diat_sampleData)
}
\description{
This sample data is a dataset used in: Nicolosi Gelis, María Mercedes; Cochero, Joaquín; Donadelli, Jorge; Gómez, Nora. 2020. "Exploring the use of nuclear alterations, motility and ecological guilds in epipelic diatoms as biomonitoring tools for water quality improvement in urban impacted lowland streams". Ecological Indicators, 110, 105951.
}
\references{
Nicolosi Gelis, María Mercedes; Cochero, Joaquín; Donadelli, Jorge; Gómez, Nora. 2020. "Exploring the use of nuclear alterations, motility and ecological guilds in epipelic diatoms as biomonitoring tools for water quality improvement in urban impacted lowland streams". Ecological Indicators, 110, 105951.
}
\keyword{datasets}
