% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diathorAll.R
\name{diathorAll}
\alias{diathorAll}
\title{This function is the master function of the package. It calculates all possible outputs from the data, and places them in the Output folder}
\usage{
diathorAll(
  species_df,
  isRelAb = F,
  maxDistTaxa = 2,
  calculateguilds = T,
  vandam = T,
  vandamReports = T,
  singleResult = T,
  exportFormat = 3,
  exportName = "Diato_results",
  plotAll = T,
  color = "#0073C2"
)
}
\arguments{
\item{species_df}{The data frame with your species data. Species as rows, Sites as columns. If empty, a dialog will prompt for a CSV file}

\item{isRelAb}{Boolean. If set to 'TRUE' it means that your species' data is the relative abundance of each species per site. If FALSE, it means that it the data corresponds to absolute densities. Default = FALSE}

\item{maxDistTaxa}{Integer. Number of characters that can differ in the species' names when compared to the internal database's name in the heuristic search. Default = 2}

\item{calculateguilds}{Boolean. If set to 'TRUE' the percentage of abundance of each diatom guild will be calculated. Default = TRUE}

\item{vandamReports}{Boolean. If set to 'TRUE' the detailed reports for the Van Dam classifications will be reported in the Output. Default = TRUE}

\item{singleResult}{Boolean. If set to 'TRUE' all results will go into a single output file. If FALSE, separate output files will be generated. Default = TRUE}

\item{exportFormat}{Integer. If = 1: A CSV (external file) will be generated with the output matrices; 2: an internal R dataframe will be generated; 3: both a CSV and an internal R dataframe are generated. Default: 3}

\item{exportName}{String. Prefix for all export files. Default: "Diato_results"}

\item{plotAll}{Boolean. If set to 'TRUE', plots will be generated for each Output in a PDF file. Default = TRUE}

\item{color}{Color code (hex). Default color for bar charts and lolipop plots. Default "#0073C2"}
}
\description{
The input file for the package is a dataframe or an external CSV file. Species should be listed as rows, with species' names in column 1.
The following columns have to be the abundance of each species (relative or absolute) in each sample (column).
The first row of the file has to contain the headers with the sample names.
If a dataframe is not specified as a parameter (species_df), the package will show a dialog box to search for the CSV file
A second dialog box will help set up an Output folder, where all outputs from the package will be exported to (dataframes, CSV files, plots in PDF)
The package downloads and installs  a wrapper for the Diat.Barcode project. Besides citing this package, the Diat.Barcode project should also be cited if the package is used, as follows:
\itemize{
\item Rimet, Frederic; Gusev, Evgenuy; Kahlert, Maria; Kelly, Martyn; Kulikovskiy, Maxim; Maltsev, Yevhen; Mann, David; Pfannkuchen, Martin; Trobajo, Rosa; Vasselon, Valentin; Zimmermann, Jonas; Bouchez, Agnès, 2018, "Diat.barcode, an open-access barcode library for diatoms", https://doi.org/10.15454/TOMBYZ, Portail Data Inra, V1
}
Sample data is taken from:
\itemize{
\item Frederic Rimet; Philippe Chaumeil; Francois Keck; Lenaïg Kermarrec; Valentin Vasselon; Maria Kahlert; Alain Franc; Agnes Bouchez. 2016. R-Syst::diatom: an open-access and curated barcode database for diatoms and freshwater monitoring. Database, 2016, 1-21.http://database.oxfordjournals.org/content/2016/baw016.full?keytype=ref&ijkey=H324uA95JzzEomz
}
\itemize{
\item Rimet F., Chaumeil P., Keck F., Kermarrec L., Vasselon V., Kahlert M., Franc A., Bouchez A., 2015. R-Syst::diatom: a barcode database for diatoms and freshwater biomonitoring - data sources and curation procedure. INRA Report, 14 pages http://dx.doi.org/10.5281/zenodo.31137
}
}
\examples{
# EXAMPLE 1:
data("environmental_data") #ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
data("species_data")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
# EXAMPLE 2: Loads sample data where species are in absolute densities
data("environmental_data_example2")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
data("species_data_example2")#ESTO HAY QUE HACER ALGUN EJEMPLO EMPAQUETADO DENTRO DEL PAQUETE
# Calculates everything
thispackage::calculate_everything(species_df)
}
