% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{epid}
\alias{epid}
\title{EPID}
\format{
A data frame with the ecological values for 1038 species
}
\source{
\url{https://www.tib.eu/en/search/id/BLCP:CN034949165/Use-of-algae-for-monitoring-rivers-in-the-Czech?cHash=fdd9e0b1bf812a31ec0f692a273cab04}
}
\usage{
data(epid)
}
\description{
Index values for diatom species included in the EPID index
}
\references{
Dell'Uomo, A. (1996). Assessment of water quality of an Apennine river as a pilot study for diatom-based monitoring of Italian watercourses. Use of algae for monitoring rivers, 65-72.
}
\keyword{datasets}
