% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{pbidw}
\alias{pbidw}
\title{PBIDW}
\format{
A data frame with the ecological values for 79 species
}
\source{
\url{https://www.researchgate.net/publication/270591536_Periphytic_diatom_index_for_assessing_the_ecological_quality_of_the_Colombian_Andean_urban_wetlands_of_Bogota}
}
\usage{
data(pbidw)
}
\description{
Index values for diatom species included in the PBIDW index
}
\references{
Castro-Roa, D., & Pinilla-Agudelo, G. (2014). Periphytic diatom index for assessing the ecological quality of the Colombian Andean urban wetlands of Bogotá. Limnetica, 33(2), 297-312.
}
\keyword{datasets}
