% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{spear}
\alias{spear}
\title{SPEAR(h)}
\format{
A data frame with the ecological values for 285 species
}
\source{
\url{https://www.sciencedirect.com/science/article/abs/pii/S1470160X18309683}
}
\usage{
data(spear)
}
\description{
Index values for diatom species included in the SPEAR(h) index
}
\references{
Wood, R. J., Mitrovic, S. M., Lim, R. P., Warne, M. S. J., Dunlop, J., & Kefford, B. J. (2019). Benthic diatoms as indicators of herbicide toxicity in rivers–A new SPEcies At Risk (SPEARherbicides) index. Ecological Indicators, 99, 203-213.
}
\keyword{datasets}
